% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVars.R
\name{createVars}
\alias{createVars}
\title{createVars}
\usage{
createVars(
  data,
  varNames = "",
  Formulas = "",
  byVars = "",
  subSets = "",
  batch = F,
  batchVars = "",
  batchVarsTail = "_new"
)
}
\description{
creating new vars
}
