% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtMnpdb.R
\name{dtMnpdb}
\alias{dtMnpdb}
\title{dtMnpdb}
\usage{
dtMnpdb(
  Host = "127.0.0.1",
  userName = "",
  passWd = "",
  dbName = "",
  tblName = "",
  query = "",
  subSets = "",
  byVars = "",
  formulas
)
}
\description{
dtMnpdb
}
