% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KA1.R
\name{KT1}
\alias{KT1}
\title{KT1}
\usage{
KT1(wd = paste0(getwd(), "/"), Encod = "utf8-8", ...)
}
\arguments{
\item{wd}{set working directory.}

\item{...}{more args used in runApp, eg. launch.browser=T}
}
\description{
launcher function
}
